refactor controller packages
