+++ src/main/java/com/ageinghippy/controller/mvc/PreparationTechniqueViewController.java
@@ -1,4 +1,4 @@
-package com.ageinghippy.controller.view;
+package com.ageinghippy.controller.mvc;
 
 import com.ageinghippy.model.dto.PreparationTechniqueDTO;
 import com.ageinghippy.service.PreparationTechniqueService;
