+++ src/main/java/com/ageinghippy/controller/rest/DishComponentController.java
@@ -1,4 +1,4 @@
-package com.ageinghippy.controller.api;
+package com.ageinghippy.controller.rest;
 
 import com.ageinghippy.model.entity.DishComponent;
 import com.ageinghippy.service.DishComponentService;
