+++ src/main/java/com/ageinghippy/controller/rest/MealController.java
@@ -1,4 +1,4 @@
-package com.ageinghippy.controller.api;
+package com.ageinghippy.controller.rest;
 
 import com.ageinghippy.model.entity.Meal;
 import com.ageinghippy.service.MealService;
