+++ src/main/java/com/ageinghippy/controller/rest/PreparationTechniqueController.java
@@ -1,4 +1,4 @@
-package com.ageinghippy.controller.api;
+package com.ageinghippy.controller.rest;
 
 import com.ageinghippy.model.dto.PreparationTechniqueDTO;
 import com.ageinghippy.service.PreparationTechniqueService;
